requirements.txt
openai
aiohttp
python-dotenv
Pillow
google-auth-oauthlib
google-auth-httplib2
google-api-python-client